Moved the Overriding of the NameConverter logic into the AbstractXJCTest so all tests can use it without having to set it
