+++ jaxb-ri/xjc/src/test/java/xjcTests/NameConverterOverrideTest.java
@@ -3,13 +3,6 @@ package xjcTests;
 import java.io.File;
 
 import org.junit.Test;
-import org.xml.sax.ErrorHandler;
-import org.xml.sax.SAXException;
-
-import com.sun.tools.xjc.BadCommandLineException;
-import com.sun.tools.xjc.Options;
-import com.sun.tools.xjc.Plugin;
-import com.sun.tools.xjc.outline.Outline;
 
 public class NameConverterOverrideTest extends AbstractXJCTest {
 
@@ -24,63 +17,7 @@ public class NameConverterOverrideTest extends AbstractXJCTest {
 				return new File(resourceDir, "Mixed.xsd");
 			}
 
-			@Override
-			protected void addOptions(Options ops) {
-				try {
-					ops.setNameConverter(new InternalNameConverter(), new LinkedPlugin());
-				} catch (BadCommandLineException e) {
-					// TODO: what is this?
-					e.printStackTrace();
-				}
-			}
-
 		});
-	}
-
-	private class LinkedPlugin extends Plugin {
-
-		@Override
-		public String getOptionName() {
-			return "linkedPlugin";
-		}
-
-		@Override
-		public String getUsage() {
-			return "EVERYWHERE";
-		}
-
-		@Override
-		public boolean run(Outline outline, Options opt, ErrorHandler errorHandler) throws SAXException {
-			// always enable for internal calls.
-			return true;
-		}
-
-	}
-
-	private class InternalNameConverter extends com.sun.xml.bind.api.impl.Standard {
-		@Override
-		public String toClassName(String s) {
-			return toMixedCaseName(toWordList(s), true);
-		}
-
-		@Override
-		public String toVariableName(String s) {
-			return s;
-		}
-
-		@Override
-		public String toInterfaceName(String token) {
-			return toClassName(token);
-		}
-
-		/**
-		 * Override the base functionality of the capitalize method to return
-		 * the same value so it matches the XSD closer
-		 */
-		@Override
-		public String capitalize(String w) {
-			return w;
-		}
 	}
 
 }
